+++ core/initializer/test_suite_constants/test_suite_launcher.go
@@ -1,0 +1,149 @@
+/*
+ * Copyright (c) 2020 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package test_suite_constants
+
+import (
+	"context"
+	"github.com/docker/go-connections/nat"
+	"github.com/kurtosis-tech/kurtosis/commons"
+	"github.com/palantir/stacktrace"
+	"net"
+	"strconv"
+)
+
+type TestsuiteContainerLauncher struct {
+	testsuiteImage string
+
+	// The log level string that will be passed as-is to the testsuite (should be meaningful to the testsuite)
+	logLevel string
+
+	// The testsuite-custom Docker environment variables that should be set in the testsuite container
+	customEnvVars map[string]string
+
+	// We'll always bind this port on the user's host machine to the testsuite, so they can do debugging
+	//  if they want
+	debuggerPort nat.Port
+}
+
+func NewTestsuiteContainerLauncher(testsuiteImage string, logLevel string, customEnvVars map[string]string, debuggerPort nat.Port) *TestsuiteContainerLauncher {
+	return &TestsuiteContainerLauncher{testsuiteImage: testsuiteImage, logLevel: logLevel, customEnvVars: customEnvVars, debuggerPort: debuggerPort}
+}
+
+/*
+Launches a new testsuite container to acquire testsuite metadata
+
+Returns: The container ID of the new container
+ */
+func (launcher TestsuiteContainerLauncher) LaunchMetadataAcquiringContainer(
+		context context.Context,
+		dockerManager *commons.DockerManager,
+		bridgeNetworkId string,
+		suiteExecutionVolume string,
+		metadataFilepathOnTestsuiteContainer string) (string, error){
+	envVars, err := launcher.generateTestSuiteEnvVars(
+		metadataFilepathOnTestsuiteContainer,
+		"", // We leave the test name blank to signify that we want test listing, not test execution
+		"", // Because we're doing test listing, not test execution, the Kurtosis API IP can be blank
+		"", // We leave the services dirpath blank because getting suite metadata doesn't require knowing this
+	)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred generating the metadata-acquiring testsuite container env vars")
+	}
+
+	resultContainerId, err := dockerManager.CreateAndStartContainer(
+		context,
+		launcher.testsuiteImage,
+		bridgeNetworkId,
+		nil,  // Nil because the bridge network will assign IPs on its own so we don't need to specify one (and won't know what IPs are already used)
+		map[nat.Port]bool{
+			launcher.debuggerPort: true,
+		},
+		nil, // Nil start command args because we expect the test suite image to be parameterized with variables
+		envVars,
+		map[string]string{},
+		map[string]string{
+			suiteExecutionVolume: TestsuiteContainerSuiteExVolMountpoint,
+		})
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred creating the test suite container to acquire testsuite metadata")
+	}
+	return resultContainerId, nil
+}
+
+/*
+Launches a new testsuite container to acquire testsuite metadata
+
+Returns: The container ID of the new container
+*/
+func (launcher TestsuiteContainerLauncher) LaunchTestRunningContainer(
+		context context.Context,
+		dockerManager *commons.DockerManager,
+		networkId string,
+		suiteExecutionVolume string,
+		testName string,
+		kurtosisApiIpStr string,
+		testsuiteContainerIp net.IP,
+		servicesRelativeDirpath string) (string, error){
+	testSuiteEnvVars, err := launcher.generateTestSuiteEnvVars(
+		"",  // We're executing a test, not getting metadata, so this should be blank
+		testName,
+		kurtosisApiIpStr,
+		servicesRelativeDirpath)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred generating the test-running testsuite container env vars")
+	}
+
+	resultContainerId, err := dockerManager.CreateAndStartContainer(
+		context,
+		launcher.testsuiteImage,
+		networkId,
+		testsuiteContainerIp,
+		map[nat.Port]bool{
+			launcher.debuggerPort: true,
+		},
+		nil,
+		testSuiteEnvVars,
+		map[string]string{},
+		map[string]string{
+			suiteExecutionVolume: TestsuiteContainerSuiteExVolMountpoint,
+		})
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred creating the test suite container to run the test")
+	}
+	return resultContainerId, nil
+}
+
+/*
+Generates the map of environment variables needed to run a test suite container
+
+NOTE: exactly one of metadata_filepath or test_name must be non-empty!
+*/
+func (launcher TestsuiteContainerLauncher) generateTestSuiteEnvVars(
+		metadataFilepath string,
+		testName string,
+		kurtosisApiIp string,
+		servicesRelativeDirpath string) (map[string]string, error){
+	debuggerPortIntStr := strconv.Itoa(launcher.debuggerPort.Int())
+	standardVars := map[string]string{
+		metadataFilepathEnvVar:        metadataFilepath,
+		testEnvVar:                    testName,
+		kurtosisApiIpEnvVar:           kurtosisApiIp,
+		servicesRelativeDirpathEnvVar: servicesRelativeDirpath,
+		logLevelEnvVar:                launcher.logLevel,
+		debuggerPort: 				   debuggerPortIntStr,
+	}
+	for key, val := range launcher.customEnvVars {
+		if _, ok := standardVars[key]; ok {
+			return nil, stacktrace.NewError(
+				"Custom test suite environment variable binding %s=%s requested, but is not allowed because key is " +
+					"already being used by Kurtosis.",
+				key,
+				val)
+		}
+		standardVars[key] = val
+	}
+	return standardVars, nil
+}