+++ core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# TBD
+# 1.2.1
 * Add a more explanatory help message to `build_and_run`
 * Correct `build_and_run.sh` to use the example microservices for kurtosis-go 1.3.0 compatibility
 * Add `docs` directory for publishing user-friendly docs (including CHANGELOG!)